combine solutions tests
