+++ unit_tests/src/agora/test_solutions.py
@@ -3,6 +3,7 @@ from decimal import Decimal
 import pytest
 
 import agora.helpers as helpers
+from agora.solutions.bfs import bfs_factory
 from agora.solutions.naive import naive_factory
 
 
@@ -19,6 +20,27 @@ def wallet() -> list[Decimal]:
         "0.01": 5,
     }
     return helpers.populate_wallet(money)
+
+
+@pytest.mark.regression
+@pytest.mark.parametrize(
+    "price, expected",
+    [
+        (
+            Decimal("15.05"),
+            [Decimal("0.05"), Decimal("20"), Decimal("5.00")],
+        ),
+        (
+            Decimal("105.06"),
+            [Decimal("100"), Decimal("0.05"), Decimal("0.01"), Decimal("10"), Decimal("5.00")],
+        ),
+        (Decimal("150.05"), [Decimal("100"), Decimal("0.05"), Decimal("100"), Decimal("50.00")]),
+    ],
+    ids=["Price: $15.05", "Price: 105.06", "Price: $150.05"],
+)
+def test_bfs(wallet, price, expected):
+    result = bfs_factory(wallet, price)()
+    assert result == (expected, len(expected)), f"Expected {expected}, but got {result}"
 
 
 @pytest.mark.regression
